FaceFusionMl/requirements.txt
fastapi[standard]==0.115.12
torch==2.6.0
torchvision==0.21.0
numpy==2.1.3
opencv-python==4.8.0.76
pillow==11.2.1
facenet-pytorch==2.5.3

